Comparisons were broken..
This fix defines a rangetuple that represents the half-open interval representing the ttcal object, i.e. [a, b) such that x is in [a, b) iff a <= x < b.

Given two such intervals, [a, b) and [c, d), they are considered: equal if there is any overlap, [a,b) < [c,d) iff b<c.